Minor bug fixes

+++ src/main/java/com/amdocs/monopoly/production/Player.java
@@ -40,7 +40,7 @@ public class Player {
 	public String getPlayerName() {
 		return name;
 	}
-	public double getPlayerBalance() {
+	public int getPlayerBalance() {
 		return balance;
 	}
